+++ src/matchers/be_some.rs
@@ -2,7 +2,7 @@ use matchers::matcher::Matcher;
 
 #[derive(Copy, Clone)]
 pub struct BeSome {
-    file_line: (&'static str, usize)
+    file_line: (&'static str, u32)
 }
 
 impl <T> Matcher<Option<T>> for BeSome {
@@ -18,12 +18,12 @@ impl <T> Matcher<Option<T>> for BeSome {
         format!("Expected {} NOT to be some but it was.", stringify!(expected))
     }
 
-    fn get_file_line(&self) -> (&'static str, usize) {
+    fn get_file_line(&self) -> (&'static str, u32) {
         self.file_line
     }
 }
 
-pub fn be_some(file_line: (&'static str, usize)) -> Box<BeSome> {
+pub fn be_some(file_line: (&'static str, u32)) -> Box<BeSome> {
     Box::new(BeSome { file_line: file_line })
 }
