+++ src/matchers/contain.rs
@@ -2,7 +2,7 @@ use matchers::matcher::Matcher;
 
 pub struct Contain<T: PartialEq> {
     value: T,
-    file_line: (&'static str, usize)
+    file_line: (&'static str, u32)
 }
 
 impl <T: PartialEq> Matcher<Vec<T>> for Contain<T> {
@@ -18,12 +18,12 @@ impl <T: PartialEq> Matcher<Vec<T>> for Contain<T> {
         format!("Expected {} NOT to contain {} but it did.", stringify!(expected), stringify!(self.value))
     }
 
-    fn get_file_line(&self) -> (&'static str, usize) {
+    fn get_file_line(&self) -> (&'static str, u32) {
         self.file_line
     }
 }
 
-pub fn contain<T: PartialEq>(value: T, file_line: (&'static str, usize)) -> Box<Contain<T>> {
+pub fn contain<T: PartialEq>(value: T, file_line: (&'static str, u32)) -> Box<Contain<T>> {
     Box::new(Contain { value: value, file_line: file_line })
 }
